Reorganize structure to accommodate non-HTML and non-Python examples
Introduced a new layer in the directory hierarchy for the programming
language. Also, pushed each example into its own subdirectory. In
addition, incorporated the file upload and job creation into the scripts
(via an imported module), so that the separate file upload script no
longer has to be manually run.